Moving SplitVar and CUIExpandableButtonState to their own files.
